renamed service interface pkg name
