+++ backend/service.go
@@ -1,4 +1,4 @@
-package model
+package backend
 
 import (
 	"context"
